pyproject.toml
[tool.poetry]
name = "ProjectZero"
version = "1.2.0"
description = "Django skeleton with most useful functionalities for new projects."
authors = ["Roman <98953084+Rokuflam@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
Django = "^5.1"
django-allauth = "^0.63"
django-extensions = "^3.2"
djangorestframework = "^3.15"
djangorestframework-simplejwt = "^5.3"
drf-spectacular = "^0.27"
uvicorn = "^0.30"
pillow = "^10.4"
psycopg2-binary = "^2.9"
python = "^3.12"
sentry-sdk = {version = "^1.45", extras = ["django"]}
whitenoise = "^6.7"
django-anymail = "^11"
celery = "^5.4"
tenacity = "^9.0"
django-redis = "^5.4"
django-cors-headers = "^4.6"
itsdangerous = "^2.2"

[tool.poetry.group.dev.dependencies]
pylint = "^3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"